Checkpoint single-use email scripts.
All of this is pretty much one-time use scripts; there are MANY
baked-in assumptions (particularly about filenames, etc.).  After
their use, they are likely only good as examples for the future --
they probably won't be useful in their current form again.

This commit also includes a Python module (written in C) that does not
write to the database, mainly because I realized that we really didn't
need to write to the database after all.  This module is still kept as
a good example of how to a) make a python module, and b) use the px
API.

Signed-off-by: Jeff Squyres <jeff@squyres.com>