sync-google-group: delete before add
Change the order of syncing to the google group: do the deletes before
the adds.  This is because we may be deleting the address of one person
and adding another address for the *same* person.  But Google may know
both of those addresses for the same Google account.  Hence, if you try
to add address B while A is already subscribed, and as far as Google is
concered, A and B are the same account, then Google will give an error
when you try to add A.

So do the deletes first -- delete A and then add B.

Signed-off-by: Jeff Squyres <jeff@squyres.com>