Summer 2018: send emails for census update
Generate and send emails for a census update.

Send an email to each active Family (i.e., send to the preferred email
address(es) for each of the Head of Household [HoH] and Spouse in the
Family) that contains links to Jotform forms:

1. A form to indicate whether the Family is still active at Epiphany,
   and if so, update their snail mail address.
2. A form (for each active Member in the family) to update basic
   census information.

Note that this script is VERY specific to the Jotform forms that are
used (i.e., there is a one-to-one correspondence to the names of the
fields in the forms to the code).

The Jotforms that the parishioners are taken to are prepopulated using
HTTP PUT semantics (i.e., https://..../?blah1=foo&blah2=bar&...).  We
did not want to send links with potentially (semi-)sensitive
information to people's email in case their email gets compromised, in
which case the attacker would get access to various PII.  Instead, we
send links with a cookie to http://api.epiphanycatholicchurch.org/.
The cookie is used to look up an HTTP PUT URL for the Jotform which
pre-populates all the relevant fields.  This cookie lookup will only
work for a fixed amount of time (e.g., if the user's email is
compromised a year later, the link will be useless).

Note that this script must be run at an IP address which is
whitelisted at Google to send emails for Epiphany (i.e., in the G
Suite dashboard, under G Suite -> GMail -> Advanced --
https://admin.google.com/epiphanycatholicchurch.org/AdminHome?hl=en#ServiceSettings/service=email&subtab=filters)
because it sends emails through Google's SMTP servers from
email-updated@epiphanycatholicchurch.org.

Signed-off-by: Jeff Squyres <jeff@squyres.com>