calendar-audit: download/save Google Calendar audit data
It turns out that Google Calendar audit data is only saved (at Google)
for 6 months.  Sometimes we need to go back farther than 6 months to
figure out what went wrong with Google Calendar issues (e.g., whether
someone accidentally deleted a Calendar event, etc.).  So have a
relatively simple Python job to run the Google Calendar audit report
and upload all the resulting JSON to a dedicated Google Team Drive.

Note that the Calendar audit data contains *ALL* Google Calendar data
for the entire G Suite domain -- even activity on individual /
personal calendars.  Hence, we default to saving this JSON in a
dedicated, highly-restricted-access Team Drive (i.e., not accessible
to the Tech Committee and not accessible to the general staff).

Note that we save the raw JSON (vs. converting it to something else --
like a spreadsheet) because according to
https://developers.google.com/admin-sdk/reports/v1/appendix/activity/calendar,
there are 30 types of Calendar audit items.  It seemed like a losing
game to try to convert them to neutral / lowest-common-denominator
items (and then keep up with changes that Google makes to these items
in the future).  Instead, just keep the raw JSON and then be able to
process it in the future (e.g., via Python) if desired.

Signed-off-by: Jeff Squyres <jeff@squyres.com>