sync-google-group: handle dup emails in a Group
A single Google account can have multiple email addresses associated
with it, e.g., A and B.  If address A is already in a Group and we try
to add B, the API call will fail with a "duplicate" result.

This "duplicate" result is not really a failure -- we can just ignore
it (i.e., we don't need to report this back in the notification
email).

Signed-off-by: Jeff Squyres <jeff@squyres.com>