export-pds: do backups at <= 12:14am
With the addition of more data exported from PDS, it was taking longer
than 5 minutes to finish, and the test for checking for "< 12:05am"
was never succeeding, and PDS data was not being backed up to Google
Drive.

So expand this test to check for "<= 12:14am".  This should be more
than enough to make sure that the backups now actually run.  Our PDS
exports run every 15 minutes, so this is still a safe test.

Signed-off-by: Jeff Squyres <jeff@squyres.com>