export-pds: new PDSChurch data directory on media-o5070
Since we're using Windows Services for Linux (WSL) on the new
media-o5070 server, we can use thei built-in network location mounter
to get the data files from \\media-o3020\pdschurch.

This will soon be replaced with a local location when we move the PDS
data files to the new media-o5070 server.

Signed-off-by: Jeff Squyres <jeff@squyres.com>