python: updates for Python 3.6.9 (vs. 3.6.2)
When moving to the new media-o5070 server, the Python for reading the
PDS SQLite3 database got a little more strict (e.g., keys were not
being found, where the key was -INT_MAX).  I'm not sure why this is
different between the old server and the new server -- the only main
difference I can find is Python 3.6.9 vs. Python 3.6.2 (the data is
the same between the old server and the new server).  Shrug.

Regardless, the change makes the Python a little tighter, and is safe
for all versions of Python (i.e., it's a good improvement,
irrespective of Python version).

Signed-off-by: Jeff Squyres <jeff@squyres.com>