Move PDS data to the "new" media server
The "new" media server has become media-o3020 (even though it is an
Optiplex 5070, not a 3020 -- but we chose to keep the old name rather
than update it everywhere that name is used).  So now the Linux Python
on this server can access the files directly through /mnt/c (not a
network share mount).

Signed-off-by: Jeff Squyres <jeff@squyres.com>