sync-google: handle a corner case in deletion
Change to delete by ID (instead of by email address) because of a weird
corner case:

- foo@example.com (a non-gmail address) is in a google group,
  but has no Google account
- later, foo@example.com visits
  https://accounts.google.com/SignupWithoutGmail and gets a
  Google account associated with that email address

In this case, Google seems to be somewhat confused: foo@example.com is
still associated with the Group, but it's the non-Google-account
foo@example.com.  But if we attempt to move that email address, it'll
try to remove the Google-account foo@example.com (and therefore fail).

So we remove by ID, and that unambiguously removes the correct member
from the Group.

Signed-off-by: Jeff Squyres <jeff@squyres.com>