BLS:  Make sure the ".exit" command is on it's on line.
In my WSL environment the .exit command throws a syntax error if it is not on a line by itself.  Media might throw an error too.
If the last DB processed creates a transaction, then it will do an 'END TRANSACTION;\n'.  The \n will put the .exit on the next line and it works.  But, if the last DB drops out from the skip/ignore cases, then the .exit will be on the same line (causing a syntax error).  In either case, the extra newline is just white space and ignore during SQL processing.

Signed-off-by: Bruce Schweinhart <bswine2@gmail.com>