Spring 2020 census code
Use a similar mechanism that we have used before:

- Mine the PDS data to send out individual emails to Families.
- A (long!) form on Jotform.com to collect all the data.  This year,
  we used a single Jotform form that had sub-pages for the Family and
  up to 7 Members.  We used Jotform conditional logic to skip Member
  sub-pages that were empty (e.g., if the PDS Family had 4 Members,
  Jotform would see that the final 3 sub-pages were empty and would
  skip them / not show them to the end user).
- Store all the data into a Google Sheet.
- Download the Sheet into a CSV and use Python scripts to mine the
  data out of it.

We noticed that a bunch of users changed their Last Name on their
Member sub-page -- including me (and I definitely did not do it on
purpose)!  We suspect that the field name we chose for at least one of
the Member sub-pages caused the user's browser to autofill the last
name field.  Might want to look a little closer on that next time.

Because the collected data was a bit "dirty", the majority of the
Member data changes were emitted to an xlsx that was uploaded to a
Google Sheet (see list_change.py), and Mary/Maria reviewed the data
manually for sanity.

Mary then extracted exactly what she wanted herself from that Sheet
and manually did imports to PDS.

We also collected *new* data this year, and turned them into PDS
keywords (e.g., typical mass attendance) or other PDS data fields
(language).  Specifically, see list_attendance.py and
list_languages.py.

Signed-off-by: Jeff Squyres <jeff@squyres.com>
Signed-off-by: Harrison Cabral <44628007+Alchemyst19@users.noreply.github.com>