Spring 2020 COVID wellness checks spreadsheets
Write out to XLSX files that can then be uploaded to Google Drive,
where they will convert -- with formatting intact -- to Google
Spreadsheets.

Note that we actually write out 2 sets of spreadsheets:

1. A bunch of names/data.
2. "missed" spreadsheets with more names/data.

This was done because the original run of this code had a bug that
accidentally skipped Families with the same last name.  That bug was
used to generate the first set of spreadsheets that were given to the
ministry members to start using.  Later, the bug was discovered, so we
decided to just generate a second set of spreadsheets that contained
all the people that were missed in the first run (vs. altering the
original spreadsheets, because the ministry had already started using
those spreadsheets).

If this code is to be used again, that should be cleaned up: don't
skip those duplicate last names and just generate one set of
spreadsheets.

Signed-off-by: Jeff Squyres <jeff@squyres.com>