2021 eStewardship
Started with a copy of the 2020-stewardship folder and significantly
progressed from there.  Here are the main, high-level changes:

1. Have one Jotform for the entire Family+Member set of submissions.
Specifically: have one Jotform with effectively 7 copies of the Member
ministry form and then 1 page with the Family financial form.  The
fields were all named deterministically so that we could
algorithmically refer to them by Member index into the Family (see
constants.py).

This significantly changed a lot of the orientation of the code:
there's only one link for the entire Family, and the reporting gets
significantly simpler (there's only one row in the output Google Sheet
for the entire Family's submission).  Note, too, that the field names
were specifically chosen to be quite short (3-4 characters long) --
see below for an exmplanation.

NOTE: We chose to have the Jotform have 7 copies of the Member
ministry data because, as of September 2020, ECC has 1 11-Member
Family and a handful of 7-Member Families.  All others Familes are 6
or less Members.

Unfortunately, using a URL with 7-Member fields' worth of data (even
with using very short field names)
apparently too much for Jotform to handle in some cases.
Specifically, when using a URL that included 7 Members worth of
pre-filled in data, the Jotform would load on an iPhone Safari and
Safari in MacOS 10.15.6 (Catalina), but when you went to finally
submit, an error would occur.  Oddly, the same thing did _not_ happen with
Chrome on MacOS, however -- Chrome/MacOS could submit 7 Members' worth
of data to Jotform just fine.  This was reported upstream to Jotform,
but they were not able to fix it in time for the 2020 eStewardship
drive:

https://www.jotform.com/answers/2534755-Prepopulate-URL-inconsistent-behaviour-on-Safari-and-Chrome-browsers-across-macOS-and-iOS

Note: the issue seemed to be with the length of the URL, not
necessarily the overall number of values in that URL.  I.e., we
specifically named the fields with very short names in order to cram
as many fields into that URL as possible without going over whatever
the limits is.

2. PDS made a new GUI element available in Church Office indicating
whether a given Family are parishioners or not.  We now obey the field
indicating that value.

Signed-off-by: Harrison Cabral <44628007+Alchemyst19@users.noreply.github.com>
Signed-off-by: Jeff Squyres <jeff@squyres.com>