eStewardship 2021: Lectors ministry
After lengthy discussions with Mary and Linda...

There are 3 ministries that are actually "combo" ministries in PDS:

- Lectors
- Acolytes
- Greeters

Each of these has 3 ministries in PDS (one for each weekend mass time)
and a fourth "extra" ministry.

The fourth ministry for Acolytes and Greeters is an "INTERESTED ONLY"
ministry.  But the fourth ministry for Lectors is a "MASTER LIST"
ministry.

Meaning:

1. You can be active in Lectors if you're active in ***any of
   the four*** Lectors ministries.
2. But you're only active in Acolytes or Greeters if you're active in
   ***any of the three mass time ministries***.

This commit adjusts the Lectors ministry to include the "master list"
ministry in the list of Lectors ministries, so that it will be
included in the "is a Member active in this ministry?" computations.

Signed-off-by: Jeff Squyres <jeff@squyres.com>