calendar-reservations: First working copy
A bot that accepts pending Google Calendar events from users with Epiphany
email domains that do not conflict with other events, and declines all other
events. It iterates through a list of all Epiphany calendars, downloading all
events. Then it checks the domain of the event creator's email, declining if
it is a non-Epiphany domain. For all events with acceptable creators, it checks
whether the event conflicts with other events on the calendar, accepting ones
that do not and declining ones that do.

Signed-off-by: Victoria Cabral <43216482+toQhom@users.noreply.github.com>