media/linux/run.sh: increase runner timeout
Sometimes Google API calls take *minutes* for no apparent reason.
Rather than fight it, just let the timeout be longer -- it won't cause
any harm, because the next iteration of the Linux python automation will be blocked
from executing by the lockfile, so it doesn't matter if this one takes
a while to run.

Signed-off-by: Jeff Squyres <jeff@squyres.com>