PDSChurch.py: normailze gmail addresses
Gmail does two special things with email addresses:

1. If the left hand side contains a ".", ignore it.
2. If the left hand size contains a "+BLAH" suffix, ignore it.

Meaning that the following email addresses all resolve to the same
Google account:

- foobar@gmail.com
- foo.bar@gmail.com
- f.o.o.b.a.r@gmail.com
- foobar+hello@gmail.com
- foobar+whazzup@gmail.com
- foo.bar+new-phone-who-dis@gmail.com

This means that we can give any of these email addresses to Google and
it will resolve them to foobar@gmail.com.

The problem is if we query a Google Group and it tells us that
"foobar@gmail.com" is a member, but we have "foo.bar@gmail.com" in
PDS.  Previously, the Python would treat these as two different email
addresses.

This commit changes it so that when we read in Google email addresses
from PDS, we normalize them:

- Remove all "." from the left hand side
- Remove all "+BLAH" suffixes from the left hand side

This way, we have the same email address in Python that Google will
report back to us from queries such as Google Group memberships, etc.

Signed-off-by: Jeff Squyres <jeff@squyres.com>