ECC PDS Census 2021.
The bulk of the code used for the ECC Census in the summer of 2021 is
in this commit.  There still remains a little more work to do in the
final processing of the data, but some of the other functionality in
this PR is needed for other projects.  Hence, I'm merging this PR with
the 90%+-done census code (and the other updates/utilities that it
has) to main, and will open a subsequent PR for the final bits of ECC
Census 2021.

Signed-off-by: Jeff Squyres <jeff@squyres.com>