Revamp Google Drive "move folder" script
Google Drive's web UI doesn't let you move a Google "My Drive" folder
into a Google Shared Drive.  Several years ago (when I barely
knew/understood Python or the Google APIs), I wrote a Python script to
traverse the source Google Drive folder and move each file to the
destination Google Shared Drive folder (called "Team Drives", back
then), creating folders at the destination as relevant.

*NOTE:* Since files in a Google Shared Drive are owned by the
organization, Google has some very specific rules about how files can
be *moved* from a My Drive folder to a Shared Drive.  Specifically:

* For files owned by Google Accounts in the destination organization,
  the owner of the original My Drive file must have write access in
  the target Shared Drive.
* For files owned by Google Accounts not in the destination
  organization, only the owner of the original My Drive file can
  *move* the file to the target Shared Drive.

As such, the orginal script attempted to *move* the file, but if that
failed, then just *copy* the file.  Copying is less desirable, but it
was good enough because the vast majority of the source files usually
fit into one of the two conditions, above (and could therefore be
*moved*).

The original script worked well enough, but had a bunch of flaws
because a) I didn't understand Python well, b) I didn't understand the
Google Drive APIs well, and c) I was pressed for time to get the job
done.  Flash forward to August of 2021: I'm still pressed for time,
but I understand Python and the Google APIs much better.  I tried to
use the original tool to do a massive (22,000+ files) My Drive -->
Google Shared Drive migration, but an error occurred partway through,
and it aborted.  This left a royal mess of half moved/copied files.

I tried updating/fixing the original script, but it was... klunky.  So
I ended up effectively re-writing the script to something new
altogether (and much simpler/better).  It does the same basic job as
the original script, but in much better ways:

* The same "try to move, but copy if necessary" philosophy is used
* Don't try to create a destination Google Shared Drive.  Instead,
  require both source and destination Google folder IDs to be
  specified by the user (i.e., the user has to create the target --
  which can even be a folder; it doesn't have to be a top-level Google
  Shared Drive).
* When copying from the source folder, only create folders at the
  target folder if a folder of the same name does not already exist.
* If a file exists at the target with the same name/kind/mime type as
  the original, assume that the file has already been copied (e.g., in
  a prior invocation) and don't copy it again.  This allows safely
  running the migration script multiple times (e.g., in case it gets
  interrupted or errors out or something) without creating N copies of
  files that can't be moved.

There's more to it than that, but that's the gist of it.

This script still isn't perfect, but hopefully it will age better than
the original.

Signed-off-by: Jeff Squyres <jeff@squyres.com>