sync-google-group.py: expand worship ministries
In the 2022 Stewardship drive, we decided to split out the following
worship ministries into sub-ministries:

1. 309 Acolytes
2. 313 Communion ministers
3. 316 Greeters
4. 318 Lectors

That is, each of these four ministries have an additional letter
appended to the number for the sub ministries.  E.g., 309A corresponds
to acolytes for the 5:30pm Saturday mass, 309B is the Sunday 9am, and
309C is the Sunday 11:30.  And so don.

And do away with "MASTER LIST" and "INTERESTED ONLY" versions of these
ministries: the A/B/C/D/etc. sub-ministries will be used to hold all
members of these ministries.  To get the "master list" of lectors, for
example, take the union of all the 318x ministries.

Signed-off-by: Jeff Squyres <jeff@squyres.com>