calendar-reservations: use "import ECC" style
This is step 1 of moving to a newer / slightly sleeker import system,
where we import ECC and ECC.Foo.  Step 2 will be to move these updated
ECC/*py files back to the top-level "python" dir and rename all the
sym links to "ecc-python-modules" to "ECC", and update all the import
statements accordingly.

Signed-off-by: Jeff Squyres <jeff@squyres.com>