Communion minister rosters: slight revamp
Updated required from the Worship ministry: create per-ministry
spreadsheets with a tab for communion ministers who are scheduleable
and another table for those who are not schedulable (because they
don't have an active mandate).  Try to emit a helpful message in the
notes column when someone is not schedulable.

Also emit a google sheet with a list of those who are about to have
their currently-active mandate expire within the next 6 months.

Signed-off-by: Jeff Squyres <jeff@squyres.com>