sync-google-group: add livestream team
Effectively revert b32e606ce, but with a new ministry name.  The
"451-Livestream Team" ministry no longer exists (because the
livestream team is not a "ministry") -- that's why b32e606ce removed
the linkage between the 451 ministry and the livestream-team@ecc
Google Group.  But we do still need to track the members of the
Livestream Team in PDS, and we want to have a Google Group for them.
So the "451-Livestream Team" ministry has been renamed to be
"WG-Livestream Team" (WG = "Working Group").

Signed-off-by: Jeff Squyres <jeff@squyres.com>