Windows Tasks Monitor - Initial Release (v01.20)
This routine will monitor tasks in the Windows task scheduler.  Specific tasks and actions for which to
monitor are specified in a JSON-formatted file specified at runtime (or using the default tasks-to-monitor
file path).  This file includes the name of the task, the state for which to monitor (currently, the last-run
state or whether the task is actively running), and a restart flag to indicate whether a task that should
run continuously should be restarted if failed.

Slack alerts will be generated if a configured task is found to not be in the specified state (or has failed
at last run, as configured).

Signed-off-by: Keith Fowler <keith.fowler.kf@gmail.com>