ECC Meraki Client Polling Initial Release (v03.80)
This routine will poll the Cisco Meraki wireless client information for the Epiphany Catholic Church network
and store the results into a SQLite database.  The client info includes both associated and non-associated
clients, which allows further analysis of "traffic" within the reach of the ECC access points, even if not
actually connected to the network.

Signed-off-by: Keith Fowler <keith.fowler.kf@gmail.com>