Process the 2021 pictorial directory Family Photos
Read the tab-separated-values (TSV) file provided by the photo vendor
and find the corresponding PDS Family from the first and last name
provided.  Produce a CSV that can be used to import the absolute
filenames of photos for Families in the "ECC Public (Z)" Google Shared
Drive, as mapped to the G: drive on an ECC Windows machine.

Signed-off-by: Jeff Squyres <jeff@squyres.com>