ricoh/report.py: add percentag columns
Add two columns for each department:

- this department's percentage of the overall B&W prints/copies in
  this time range
- this department's percentage of the overall color prints/copies in
  this time range

We add these columns because of a discovery we made this month:
Duplicator Sales (our Ricoh maintenance vendor) charges us monthly
based on the Meter History of the Ricoh.  This Meter History is from
the "Counter" report on the Ricoh, and it represents the total usage
since the beginning of the life of this printer.

    Sidenote: Duplicator Sales records the Meter History counter
    values on the 8th of each month (it's an automated process).

Every night, ECC downloads the "Counter by User" CSV, which is a
point-in-time reporting of all kinds of statistics (including total
B&W and total color) for each user (i.e., ECC department).  We record
that CSV data in an SQLite database.

However, while the "Counter" Ricoh report is involatile, the "Counter
by User" report can actually have its per-user numbers reset to 0,
such as when we deleted the Worship department user on Dec 15 2021,
and then re-created the user with the same department ID.  This
effectively zeroed out the usage for the Worship department.  Hence,
if you add up all the "BW" and "Color" totals from the "Counter by
User" Ricoh report, they do *not* add up to the totals from the
"Counter" report, because we have deleted some user data (there's
additional discrepancies because we capture the "Counter by User" CSV
at 12:01am each day, but we deleted/re-added the Worship department at
some point later in the day, thereby losing however many BW/Color were
made after midnight and before the delete/re-add).

As such, since our reports won't exactly match the totals from
Duplicator Sales, the total number of BW/Color printed per department
for a given date range (e.g., one month) -- as long as that date range
does not include Dec 15, 2021 -- will be accurage.  Hence, we just
show each department's percentage of BW and Color compared to the
total BW and Color.  ECC can then use those percentages to compute
chargebacks to the departments.

Signed-off-by: Jeff Squyres <jeff@squyres.com>