sync-google-groups: update
Add top-level ministries for 309, 313, 316, and 318.

As of August 2022, it has been decided that we are removing the
A/B/C/D ministries altogether.  Data has been sent to the parish admin
assistant to update PDS such that all Members who are in the A/B/C/D
sub-ministries are *also* in the no-letter ministries.  This will
allow us to remove the A/B/C/D ministries from PDS altogether.

During this transition, sync the Google Groups to the A/B/C/D
ministries *and* the no-letter ministry.  A future commit will remove
syncing against the A/B/C/D ministries.

Signed-off-by: Jeff Squyres <jeff@squyres.com>