sync-google-group: remove A/B/C/D sub-ministries
For each of 309, 313, 316, and 318, remove all the A/B/C/D
sub-ministries, and remove the "INTERESTED ONLY" / "MASTER LIST"
designations on the corresponding no-letter ministries.

This follows the decision to eliminate the A/B/C/D sub-ministries and
just have all Members in the corresponding no-letter ministries.

(this is because we are now tracking which times Members want to serve
in Ministry Scheduler Pro, not PDS)

Signed-off-by: Jeff Squyres <jeff@squyres.com>