run-all.py: workaround Oct 2022 WS19/WSL1 errors
In Oct 2022, about once a day, some run.sh script started hanging
indefinitely.  Specifically, there will be a child "awk" process of
export-pds-into-sqlite/run.sh, and the awk process will complete and
go into a zombie state. However, the parent run.sh script apparently
does not recognize that the child is now zombied and needs to be
reaped.  It sits, forever. This means that the subsequent invocation
of run-all.py will fail to get the lockfile.  Eventually, it will
start sending error messages to Slack because the lockfile from the
still-hung run-all.py will become too old.

This parent-does-not-reap-a-child-zombie behavior very much feels like
a bug in WS2019/WSL1.  Indeed, I do not even see anywhere where the
code from this tree invokes "awk" -- that's a bit of a mystery in
itself.

For the moment, simply remove a lockfile that is "too old" and
continue on.

Signed-off-by: Jeff Squyres <jeff@squyres.com>