gsheet-driven-google-group.py: fix email wording
A trivial missing computation accidentally sent the wrong wording to
the addressees who are being removed from the Google Group.  Add the
missing line to properly compute the list of email addresses that is
insertted into the email body text indicating who will no longer be
receiving the Group emails.

Signed-off-by: Jeff Squyres <jeff@squyres.com>