create-ministry-rosters: have PS rosters write to old PDS files
The old Google Sheets created for the PDS rosters have been linked and
shared to a lot of places.  Since the PDS roster creation script is
not being run any more, just update the PS roster creation script to
point to the same document IDs as the old PDS roster creation script
so that all the files remain the same -- they're just getting filled
up with PS data now (not PDS data).  This means that we don't have to
re-share, re-link files -- all the old shares and links will still work.

Signed-off-by: Jeff Squyres <jeff@squyres.com>