ricoh: automate monthly report email
We previously developed monthyl, quarterly, and annual Ricoh reports.
However, it turns out that this is not what is needed.  All we need is
a monthly report.  So return to the monthly report logic, and add the
functionality of automatically uploading it to Google and sending an
email when the upload is complete.

Signed-off-by: Jeff Squyres <jeff@squyres.com>